tests/fixtures/ai_rag_app/requirements.txt
openai>=1.0.0
pinecone-client>=3.0.0
langchain>=0.2.0
llama-index>=0.10.0
chromadb>=0.4.0
fastapi>=0.115.0
pydantic>=2.0.0
